+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+* `PackageRepository.download` now has more specific return type:
+  `Future<Stream<List<int>>>`.
+
 ## 0.1.2
 
 * Add support for generic exceptions raised e.g. due to `pubspec.yaml`